requirements.txt
Flask_SQLAlchemy==2.3.2
Flask==1.0.2
geopy==1.20.0
googlemaps==3.1.3
gmplot==1.2.0
Flask_WTF==0.14.2
WTForms==2.2.1
python_google_places==1.4.1
protobuf==3.14.0
